+++ src/console/auto_switcher.ts
