+++ src/console/connection_scanner.ts
