+++ src/dolphin/install/extract_dmg.ts
