+++ src/dolphin/setup.ts
@@ -3,7 +3,7 @@ import log from "electron-log";
 import * as fs from "fs-extra";
 import { isEqual } from "lodash";
 import path from "path";
-import { fileExists } from "utils/fileExists";
+import { fileExists } from "utils/file_exists";
 
 import {
   ipc_checkDesktopAppDolphin,
