+++ src/main/network_diagnostics.ts
