+++ src/renderer/pages/settings/dolphin_settings/gecko_codes/manage_codes/manage_codes.container.tsx
@@ -1,5 +1,5 @@
-import type { GeckoCode } from "@dolphin/config/geckoCode";
-import { geckoCodeToString } from "@dolphin/config/geckoCode";
+import type { GeckoCode } from "@dolphin/config/gecko_code";
+import { geckoCodeToString } from "@dolphin/config/gecko_code";
 
 import { useToasts } from "@/lib/hooks/use_toasts";
